Add idapython script to find rtti tables
port the C# rtti finding code, to python, then change the code from a
scan memory A to B of vtables, make it find the references on the object
descriptor strings.
This allows finding all references, but this step gets broken by the
multi inheritance tables...

But ignoring that it is already a lot better than the old two step
method, and I can next put the data type formatting code, in, at least
this all makes porting from one camera model to the next easier.